requirements.txt
azure-functions==1.17.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-mgmt-storage==21.1.0
azure-core==1.29.5